nem: layout refactoring and fixes
